pyproject.toml
[project]
name = "writerassistant"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pygame>=2.6.1",
    "pygame-gui>=0.6.14",
    "torch>=2.8.0",
    "transformers>=4.57.0",
    "python-dotenv",
    "pytest",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
]
